requirements.txt
langchain
langchain-groq
langchain-community
langgraph
duckduckgo-search
python-dotenv
groq